+++ src/main/java/cn/edu/tongji/healper/service/HistoryService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface HistoryService {
 
-    Integer addConsultHistory(Integer clientId, Integer consultantId, Integer expense, String status);
+    Integer addConsultHistory(Integer clientId, Integer consultantId, Double expense, String status);
 
     String findQrCodeByHistoryId(Integer historyId);
